+++ src/@types/data/index.d.ts
@@ -3,3 +3,8 @@ export type NetworkBoundResult = {
   message: string;
   data: any | null;
 };
+
+export type OfflineResult = {
+  success: boolean;
+  data: any | null;
+};