+++ src/extension/commands.ts
@@ -13,20 +13,46 @@ import { getMimicHex7, validateHex6 } from "../color";
 import {
   getAllAuxiliaryThemeRegistryIndexes as getAuxiliaryThemeRegistryIndexes,
   prepareAuxiliaryTheme,
+  prepareAuxiliaryThemeOffline,
   prepareAuxiliaryThemeRegistries,
+  prepareAuxiliaryThemeRegistriesOffline,
 } from "../data";
 import { getAuxiliaryThemeId } from "../data/helpers";
-import { mimic3Info as darkMimic } from "../modern/variants/dark/modern";
-import { mimic1Info as lightMimic } from "../modern/variants/light/modern";
+import {
+  mimic3Info as darkMimic,
+  palette as darkPalette,
+} from "../modern/variants/dark/modern";
+import {
+  mimic1Info as lightMimic,
+  palette as lightPalette,
+} from "../modern/variants/light/modern";
 import { authenticate } from "./authentication";
 import { NOTIFICATION_SIGNATURE } from "./constants";
 import { toggleFirstLetterCase } from "./helpers";
-import { generateAdaptiveModeIcons } from "./icons";
+import { generateAccentColorIcons, generateAdaptiveModeIcons } from "./icons";
 import { showProgressNotification } from "./notifications";
-import { setIsConfiguredFromCommand } from "./sharedState";
-import { getConfig, updateConfig } from "./utils";
+import { getIsOfflineMode, setIsConfiguredFromCommand } from "./sharedState";
+import { checkInternetConnection, getConfig, updateConfig } from "./utils";
+
+export const authenticateCommand = async () => {
+  showProgressNotification(
+    `${NOTIFICATION_SIGNATURE} Authenticating...`,
+    async () => {
+      const result = await authenticate(true);
+      if (!result.success) {
+        if (result.message) {
+          window.showErrorMessage(result.message);
+        }
+      } else {
+        window.showInformationMessage("Authenticated successfully!");
+      }
+    },
+  );
+};
 
 export const configureCommand = async () => {
+  await checkInternetConnection();
+  const isOfflineMode = getIsOfflineMode();
   //
   const variantLabel = await getVariantLabel();
   if (!variantLabel) {
@@ -37,6 +63,7 @@ export const configureCommand = async () => {
   const auxiliaryUiThemeExtension = await getAuxiliaryThemeExtension(
     variant,
     "ui",
+    isOfflineMode,
   );
   if (!auxiliaryUiThemeExtension) {
     return;
@@ -47,6 +74,7 @@ export const configureCommand = async () => {
       variant,
       auxiliaryUiThemeExtension,
       "ui",
+      isOfflineMode,
     );
     if (!auxiliaryUiThemeCandidate) {
       return;
@@ -63,9 +91,21 @@ export const configureCommand = async () => {
       designLabel.replace(/\$\(.*\) /g, ""),
     ) as Design;
   }
-  const accentColor = !auxiliaryUiTheme ? await getAccentColor(variant) : null;
-  if (!accentColor && !auxiliaryUiTheme) {
+  //
+  const accentColorFromPalette = !auxiliaryUiTheme
+    ? await getAccentColorFromPalette(variant)
+    : null;
+  if (!accentColorFromPalette && !auxiliaryUiTheme) {
     return;
+  }
+  let accentColor: string | null | undefined;
+  if (accentColorFromPalette !== "_") {
+    accentColor = accentColorFromPalette;
+  } else {
+    accentColor = !auxiliaryUiTheme ? await getAccentColor(variant) : null;
+    if (!accentColor && !auxiliaryUiTheme) {
+      return;
+    }
   }
   //
   const adaptiveModeLabel = !auxiliaryUiTheme
@@ -82,6 +122,7 @@ export const configureCommand = async () => {
   const auxiliaryCodeThemeExtension = await getAuxiliaryThemeExtension(
     variant,
     "code",
+    isOfflineMode,
   );
   if (!auxiliaryCodeThemeExtension) {
     return;
@@ -92,6 +133,7 @@ export const configureCommand = async () => {
       variant,
       auxiliaryCodeThemeExtension,
       "code",
+      isOfflineMode,
     );
     if (!auxiliaryCodeThemeCandidate) {
       return;
@@ -114,22 +156,6 @@ export const configureCommand = async () => {
       `Codemos Modern (${toggleFirstLetterCase(variant)})`,
       ConfigurationTarget.Global,
     );
-};
-
-export const authenticateCommand = async () => {
-  showProgressNotification(
-    `${NOTIFICATION_SIGNATURE} Authenticating...`,
-    async () => {
-      const result = await authenticate(true);
-      if (!result.success) {
-        if (result.message) {
-          window.showErrorMessage(result.message);
-        }
-      } else {
-        window.showInformationMessage("Authenticated successfully!");
-      }
-    },
-  );
 };
 
 const getVariantLabel = async () => {
@@ -215,6 +241,271 @@ const getDesign = async (variant: Variant) => {
   return design.label;
 };
 
+const getAccentColorFromPalette = async (variant: Variant) => {
+  const quickPick = window.createQuickPick();
+  quickPick.title = `Codemos Modern (${toggleFirstLetterCase(
+    variant,
+  )}): UI Theme`;
+  quickPick.placeholder = "Select an accent color";
+  quickPick.ignoreFocusOut = true;
+  quickPick.busy = true;
+  quickPick.show();
+  prepareAccentColorIcons(variant);
+  quickPick.items = [
+    {
+      label: "Custom",
+      kind: QuickPickItemKind.Separator,
+    },
+    {
+      label: "Custom",
+      description: "Accent color",
+      detail: "A color you specify",
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_custom.svg`,
+      ),
+    },
+    {
+      label: "Palette",
+      kind: QuickPickItemKind.Separator,
+    },
+    {
+      label: "Brown",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "antique brass" : "mud"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_brown.svg`,
+      ),
+    },
+    {
+      label: "Red",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "burnt sienna" : "guardsman red"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_red.svg`,
+      ),
+    },
+    {
+      label: "Orange",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "raw sienna" : "chelsea gem"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_orange.svg`,
+      ),
+    },
+    {
+      label: "Yellow",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "roti" : "olive"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_yellow.svg`,
+      ),
+    },
+    {
+      label: "Green",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "apple" : "camarone"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_green.svg`,
+      ),
+    },
+    {
+      label: "Mint",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "keppel" : "blue stone"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_mint.svg`,
+      ),
+    },
+    {
+      label: "Blue",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "Shakespeare" : "science blue"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_blue.svg`,
+      ),
+    },
+    {
+      label: "Purple",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "Portage" : "electric violet"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_purple.svg`,
+      ),
+    },
+    {
+      label: "Pink",
+      description: "Accent color",
+      detail: `The color of ${
+        variant === "dark" ? "brilliant rose" : "flirt"
+      } from the Modern's palette`,
+      iconPath: Uri.file(
+        `${__dirname}/../../res/icons/${variant}/accent_pink.svg`,
+      ),
+    },
+  ];
+  quickPick.busy = false;
+  return await new Promise<string | undefined>((resolve) => {
+    quickPick.onDidAccept(() => {
+      const selectedAccentColorLabel = quickPick.selectedItems[0].label;
+      quickPick.dispose();
+      if (selectedAccentColorLabel) {
+        switch (variant) {
+          case "dark":
+            switch (selectedAccentColorLabel) {
+              case "Custom":
+                return resolve("_");
+              case "Brown":
+                return resolve(darkPalette.basic.def.brown);
+              case "Red":
+                return resolve(darkPalette.basic.def.red);
+              case "Orange":
+                return resolve(darkPalette.basic.def.orange);
+              case "Yellow":
+                return resolve(darkPalette.basic.def.yellow);
+              case "Green":
+                return resolve(darkPalette.basic.def.green);
+              case "Mint":
+                return resolve(darkPalette.basic.def.mint);
+              case "Blue":
+                return resolve(darkPalette.basic.def.blue);
+              case "Purple":
+                return resolve(darkPalette.basic.def.purple);
+              case "Pink":
+                return resolve(darkPalette.basic.def.pink);
+              default:
+                return resolve(undefined);
+            }
+          case "light":
+            switch (selectedAccentColorLabel) {
+              case "Custom":
+                return resolve("_");
+              case "Brown":
+                return resolve(lightPalette.basic.def.brown);
+              case "Red":
+                return resolve(lightPalette.basic.def.red);
+              case "Orange":
+                return resolve(lightPalette.basic.def.orange);
+              case "Yellow":
+                return resolve(lightPalette.basic.def.yellow);
+              case "Green":
+                return resolve(lightPalette.basic.def.green);
+              case "Mint":
+                return resolve(lightPalette.basic.def.mint);
+              case "Blue":
+                return resolve(lightPalette.basic.def.blue);
+              case "Purple":
+                return resolve(lightPalette.basic.def.purple);
+              case "Pink":
+                return resolve(lightPalette.basic.def.pink);
+              default:
+                return resolve(undefined);
+            }
+          default:
+            return resolve(undefined);
+        }
+      } else {
+        return resolve(undefined);
+      }
+    });
+  });
+};
+
+const prepareAccentColorIcons = (variant: Variant) => {
+  let brown: string;
+  let red: string;
+  let orange: string;
+  let yellow: string;
+  let green: string;
+  let mint: string;
+  let blue: string;
+  let purple: string;
+  let pink: string;
+  if (variant === "dark") {
+    brown = darkPalette.basic.def.brown;
+    red = darkPalette.basic.def.red;
+    orange = darkPalette.basic.def.orange;
+    yellow = darkPalette.basic.def.yellow;
+    green = darkPalette.basic.def.green;
+    mint = darkPalette.basic.def.mint;
+    blue = darkPalette.basic.def.blue;
+    purple = darkPalette.basic.def.purple;
+    pink = darkPalette.basic.def.pink;
+  } else {
+    brown = lightPalette.basic.def.brown;
+    red = lightPalette.basic.def.red;
+    orange = lightPalette.basic.def.orange;
+    yellow = lightPalette.basic.def.yellow;
+    green = lightPalette.basic.def.green;
+    mint = lightPalette.basic.def.mint;
+    blue = lightPalette.basic.def.blue;
+    purple = lightPalette.basic.def.purple;
+    pink = lightPalette.basic.def.pink;
+  }
+  switch (window.activeColorTheme.kind) {
+    case ColorThemeKind.Dark || ColorThemeKind.HighContrast:
+      generateAccentColorIcons(
+        variant,
+        true,
+        brown,
+        red,
+        orange,
+        yellow,
+        green,
+        mint,
+        blue,
+        purple,
+        pink,
+      );
+      break;
+    case ColorThemeKind.Light || ColorThemeKind.HighContrastLight:
+      generateAccentColorIcons(
+        variant,
+        false,
+        brown,
+        red,
+        orange,
+        yellow,
+        green,
+        mint,
+        blue,
+        purple,
+        pink,
+      );
+      break;
+    default:
+      generateAccentColorIcons(
+        variant,
+        true,
+        brown,
+        red,
+        orange,
+        yellow,
+        green,
+        mint,
+        blue,
+        purple,
+        pink,
+      );
+      break;
+  }
+};
+
 const getAccentColor = async (variant: Variant) => {
   const accentColorHex7 = await window.showInputBox({
     title: `Codemos Modern (${toggleFirstLetterCase(variant)}): UI Theme`,
@@ -239,7 +530,7 @@ const getAdaptiveModeLabel = async (
   variant: Variant,
   accentColorHex7: string,
 ) => {
-  const quickPick = await window.createQuickPick();
+  const quickPick = window.createQuickPick();
   quickPick.title = `Codemos Modern (${toggleFirstLetterCase(
     variant,
   )}): UI Theme`;
@@ -247,7 +538,7 @@ const getAdaptiveModeLabel = async (
   quickPick.ignoreFocusOut = true;
   quickPick.busy = true;
   quickPick.show();
-  prepareIcons(variant, accentColorHex7);
+  prepareAdaptiveModeIcons(variant, accentColorHex7);
   quickPick.items = [
     {
       label: "None",
@@ -296,7 +587,10 @@ const getAdaptiveModeLabel = async (
   });
 };
 
-const prepareIcons = (variant: Variant, accentColorHex7: string) => {
+const prepareAdaptiveModeIcons = (
+  variant: Variant,
+  accentColorHex7: string,
+) => {
   let noneModeColor: string;
   let gentleModeColor: string;
   let moderateModeColor: string;
@@ -371,7 +665,7 @@ const prepareIcons = (variant: Variant, accentColorHex7: string) => {
   }
 };
 
-interface AuxiliaryThemeExtensionAsQuickPickItem extends QuickPickItem {
+interface AuxiliaryThemeQPI extends QuickPickItem {
   auxiliaryThemeId: string | undefined;
   label: string;
   description?: string;
@@ -383,6 +677,7 @@ interface AuxiliaryThemeExtensionAsQuickPickItem extends QuickPickItem {
 const getAuxiliaryThemeExtension = async (
   variant: Variant,
   kind: "ui" | "code",
+  isOfflineMode: boolean,
 ) => {
   const kindLabel = kind === "ui" ? "UI" : "Code";
   const quickPick = window.createQuickPick();
@@ -394,9 +689,12 @@ const getAuxiliaryThemeExtension = async (
   quickPick.busy = true;
   quickPick.show();
   const auxiliaryThemeRegistries = getConfig().auxiliaryThemeRegistries;
-  const success = await prepareAuxiliaryThemeRegistries(
-    auxiliaryThemeRegistries,
-  );
+  let success: boolean;
+  if (isOfflineMode) {
+    success = prepareAuxiliaryThemeRegistriesOffline(auxiliaryThemeRegistries);
+  } else {
+    success = await prepareAuxiliaryThemeRegistries(auxiliaryThemeRegistries);
+  }
   if (!success) {
     quickPick.dispose();
     return undefined;
@@ -404,26 +702,23 @@ const getAuxiliaryThemeExtension = async (
   const auxiliaryThemeRegistryIndexesWithId = getAuxiliaryThemeRegistryIndexes(
     auxiliaryThemeRegistries,
   );
-  const auxiliaryThemeAsQuickPickItems: AuxiliaryThemeExtensionAsQuickPickItem[] =
-    [];
-  auxiliaryThemeAsQuickPickItems.push({
+  const auxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
+  auxiliaryThemesQPI.push({
     auxiliaryThemeId: undefined,
     label: "Bundled",
     kind: QuickPickItemKind.Separator,
   });
-  auxiliaryThemeAsQuickPickItems.push({
+  auxiliaryThemesQPI.push({
     auxiliaryThemeId: "_",
     label: "$(library) Codemos Modern",
     description: "Codemos-Inc/Codemos-Modern $(verified-filled)",
     detail: `$(organization) Codemos • $(compass) Bundled • $(law) MIT`,
     alwaysShow: true,
   });
-  auxiliaryThemeAsQuickPickItems.push({
-    auxiliaryThemeId: undefined,
-    label: "From registries",
-    kind: QuickPickItemKind.Separator,
-  });
+  const installedAuxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
+  const availableAuxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
   for (const auxiliaryThemeRegistryIndexWithId of auxiliaryThemeRegistryIndexesWithId) {
+    // 🟢 Optimize filtering algorithm
     const filteredAuxiliaryThemes =
       auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryIndex.themes[
         variant
@@ -436,16 +731,33 @@ const getAuxiliaryThemeExtension = async (
           )
         );
       });
+    for (const filteredAuxiliaryTheme of filteredAuxiliaryThemes) {
+      const installedOption =
+        auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryIndex.themes[
+          variant
+        ].find((auxiliaryTheme) => {
+          return (
+            auxiliaryTheme.extension === filteredAuxiliaryTheme.extension &&
+            auxiliaryTheme.installed
+          );
+        });
+      if (installedOption) {
+        const index = filteredAuxiliaryThemes.findIndex((auxiliaryTheme) => {
+          return auxiliaryTheme === filteredAuxiliaryTheme;
+        });
+        filteredAuxiliaryThemes[index] = installedOption;
+      }
+    }
+    const isVerifiedOwner = verifiedOwners.find((verifiedOwner) => {
+      return (
+        verifiedOwner ===
+        auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner.toLowerCase()
+      );
+    })
+      ? true
+      : false;
     for (const auxiliaryTheme of filteredAuxiliaryThemes) {
-      const isVerifiedOwner = verifiedOwners.find((verifiedOwner) => {
-        return (
-          verifiedOwner ===
-          auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner.toLowerCase()
-        );
-      })
-        ? true
-        : false;
-      auxiliaryThemeAsQuickPickItems.push({
+      const auxiliaryThemeExtensionQPI: AuxiliaryThemeQPI = {
         auxiliaryThemeId: `${auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner}/${auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.repo}/${auxiliaryTheme.publisher}/${auxiliaryTheme.extension}/${auxiliaryTheme.theme}`,
         label: `$(extensions) ${auxiliaryTheme.extension}`,
         description: `${
@@ -454,19 +766,41 @@ const getAuxiliaryThemeExtension = async (
           isVerifiedOwner ? "$(verified-filled)" : "$(unverified)"
         }`,
         detail: `$(organization) ${auxiliaryTheme.publisher} • $(compass) ${auxiliaryTheme.origin} • $(law) ${auxiliaryTheme.license}`,
-      });
+      };
+      if (auxiliaryTheme.installed) {
+        installedAuxiliaryThemesQPI.push(auxiliaryThemeExtensionQPI);
+      } else {
+        if (!isOfflineMode) {
+          availableAuxiliaryThemesQPI.push(auxiliaryThemeExtensionQPI);
+        }
+      }
     }
   }
+  if (installedAuxiliaryThemesQPI.length > 0) {
+    auxiliaryThemesQPI.push({
+      auxiliaryThemeId: undefined,
+      label: "Installed",
+      kind: QuickPickItemKind.Separator,
+    });
+    auxiliaryThemesQPI.push(...installedAuxiliaryThemesQPI);
+  }
+  if (availableAuxiliaryThemesQPI.length > 0) {
+    auxiliaryThemesQPI.push({
+      auxiliaryThemeId: undefined,
+      label: "Available",
+      kind: QuickPickItemKind.Separator,
+    });
+    auxiliaryThemesQPI.push(...availableAuxiliaryThemesQPI);
+  }
   quickPick.busy = false;
-  quickPick.items = auxiliaryThemeAsQuickPickItems;
+  quickPick.items = auxiliaryThemesQPI;
   return await new Promise<string | undefined>((resolve) => {
     quickPick.onDidAccept(() => {
       const selectedAuxiliaryTheme = quickPick.selectedItems[0];
       quickPick.dispose();
       if (selectedAuxiliaryTheme) {
         return resolve(
-          (selectedAuxiliaryTheme as AuxiliaryThemeExtensionAsQuickPickItem)
-            .auxiliaryThemeId,
+          (selectedAuxiliaryTheme as AuxiliaryThemeQPI).auxiliaryThemeId,
         );
       } else {
         return resolve(undefined);
@@ -479,6 +813,7 @@ const getAuxiliaryTheme = async (
   variant: Variant,
   auxiliaryThemeExtension: string,
   kind: "ui" | "code",
+  isOfflineMode: boolean,
 ) => {
   const kindLabel = kind === "ui" ? "UI" : "Code";
   const quickPick = window.createQuickPick();
@@ -495,8 +830,9 @@ const getAuxiliaryTheme = async (
   const auxiliaryThemeRegistryIndexesWithId = getAuxiliaryThemeRegistryIndexes(
     auxiliaryThemeRegistries,
   );
-  const auxiliaryThemeAsQuickPickItems: AuxiliaryThemeExtensionAsQuickPickItem[] =
-    [];
+  const auxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
+  const installedAuxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
+  const availableAuxiliaryThemesQPI: AuxiliaryThemeQPI[] = [];
   const auxiliaryThemeId = getAuxiliaryThemeId(auxiliaryThemeExtension);
   for (const auxiliaryThemeRegistryIndexWithId of auxiliaryThemeRegistryIndexesWithId) {
     const filteredAuxiliaryThemes =
@@ -508,16 +844,16 @@ const getAuxiliaryTheme = async (
           auxiliaryTheme.extension === auxiliaryThemeId.extension
         );
       });
+    const isVerifiedOwner = verifiedOwners.find((verifiedOwner) => {
+      return (
+        verifiedOwner ===
+        auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner.toLowerCase()
+      );
+    })
+      ? true
+      : false;
     for (const auxiliaryTheme of filteredAuxiliaryThemes) {
-      const isVerifiedOwner = verifiedOwners.find((verifiedOwner) => {
-        return (
-          verifiedOwner ===
-          auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner.toLowerCase()
-        );
-      })
-        ? true
-        : false;
-      auxiliaryThemeAsQuickPickItems.push({
+      const auxiliaryThemeQPI: AuxiliaryThemeQPI = {
         auxiliaryThemeId: `${auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.owner}/${auxiliaryThemeRegistryIndexWithId.auxiliaryThemeRegistryId.repo}/${auxiliaryTheme.publisher}/${auxiliaryTheme.extension}/${auxiliaryTheme.theme}`,
         label: `$(symbol-color) ${auxiliaryTheme.theme}`,
         description: `${
@@ -528,24 +864,57 @@ const getAuxiliaryTheme = async (
         detail: `$(organization) ${auxiliaryTheme.publisher} • $(extensions) ${
           auxiliaryTheme.extension
         } • $(color-mode) ${toggleFirstLetterCase(variant)}`,
-      });
+      };
+      if (auxiliaryTheme.installed) {
+        installedAuxiliaryThemesQPI.push(auxiliaryThemeQPI);
+      } else {
+        if (!isOfflineMode) {
+          availableAuxiliaryThemesQPI.push(auxiliaryThemeQPI);
+        }
+      }
     }
   }
+  if (installedAuxiliaryThemesQPI.length > 0) {
+    auxiliaryThemesQPI.push({
+      auxiliaryThemeId: undefined,
+      label: "Installed",
+      kind: QuickPickItemKind.Separator,
+    });
+    auxiliaryThemesQPI.push(...installedAuxiliaryThemesQPI);
+  }
+  if (availableAuxiliaryThemesQPI.length > 0) {
+    auxiliaryThemesQPI.push({
+      auxiliaryThemeId: undefined,
+      label: "Available",
+      kind: QuickPickItemKind.Separator,
+    });
+    auxiliaryThemesQPI.push(...availableAuxiliaryThemesQPI);
+  }
   quickPick.busy = false;
-  quickPick.items = auxiliaryThemeAsQuickPickItems;
+  quickPick.items = auxiliaryThemesQPI;
   return await new Promise<string | undefined>((resolve) => {
     quickPick.onDidAccept(async () => {
       const selectedAuxiliaryTheme = quickPick.selectedItems[0];
       quickPick.dispose();
       if (selectedAuxiliaryTheme) {
         const selectedAuxiliaryThemeId = (
-          selectedAuxiliaryTheme as AuxiliaryThemeExtensionAsQuickPickItem
+          selectedAuxiliaryTheme as AuxiliaryThemeQPI
         ).auxiliaryThemeId;
         if (selectedAuxiliaryThemeId) {
-          const success = await prepareAuxiliaryTheme(
-            selectedAuxiliaryThemeId,
-            variant,
-          );
+          let success: boolean;
+          if (isOfflineMode) {
+            success = await prepareAuxiliaryThemeOffline(
+              auxiliaryThemeRegistries,
+              selectedAuxiliaryThemeId,
+              variant,
+            );
+          } else {
+            success = await prepareAuxiliaryTheme(
+              auxiliaryThemeRegistries,
+              selectedAuxiliaryThemeId,
+              variant,
+            );
+          }
           if (!success) {
             return resolve(undefined);
           }