+++ src/extension/messages.ts
@@ -1,0 +1,3 @@
+export const offlineModeActive = (): string => {
+  return `No internet connection! Offline mode is active.`;
+};