+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Nothing yet.
 
+## [2.1.0] - 2023-08-25 - Offline availability and color presets 📡
+
+### Added
+
+- Offline availability. Modern will now operate when the users are offline, allowing them to pick amongst installed auxiliary themes.
+- Color presets. Users can now choose an accent color from Modern's excellent palette.
+
+### Changed
+
+- Visualization of auxiliary themes and extensions in command `codemosModern.configure`, indicating which of them are installed.
+
+### Fixed
+
+- Wrong scoping of theme change resulting all sorts of headaches.
+
 ## [2.0.1] - 2023-08-17 - Scoping hotfix 🛠️
 
 ### Fixed
@@ -175,7 +190,9 @@ Nothing yet.
 
 - Initial release
 
-[unreleased]: https://github.com/Codemos-Inc/codemos-modern/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/Codemos-Inc/codemos-modern/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/Codemos-Inc/codemos-modern/compare/v2.0.1...v2.1.0
+[2.0.1]: https://github.com/Codemos-Inc/codemos-modern/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.2.2...v2.0.0
 [1.2.2]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/Codemos-Inc/codemos-modern/compare/v1.2.0...v1.2.1